refactor: move the oauth2 client-credentials auth into the common package
Consolidate on the hardened finops variant: expiry short-circuit,
double-checked locking around refreshes, and credential-sanitized URLs
in logs. TLS verification becomes a constructor argument; the
settings-bound factory and connection check stay in per-agent shims.
sre-agent picks up the hardened behavior; portal-assistant has no
oauth flow and is untouched.
