+++ agents/common/pyproject.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 description = "Shared infrastructure for the OpenChoreo Python agent services"
 requires-python = ">=3.14"
 dependencies = [
+    "authlib>=1.6.8",
     "fastapi>=0.139.2",
     "httpx>=0.23.0",
     "jinja2>=3.1.0",
